Progress
Started and working on cacheContents() function and the process of how I am going to cache all my files/folders. Working on making my unordered_multimap to work where the key : value is boolean : Pathing() parent class. Where the boolean will set to True if a folder, and false if a file. And the value uses the parent class to set it to a derived child class either File() or Folder(). Trying to figure out how to access the derived classes variables while looping through the map