previousPath start process
start the process of keeping the previous path stored and sharing it between instances of caching Contents on demand while the user navigates the code. Also working on making a tree or linked list to store a data structure representing a directory and all the folders and their child folders. Figuring out if I want to do it this way and use a tree and just navigate it this way instead of caching on demand.